% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_normalization.R
\name{back_normalization}
\alias{back_normalization}
\title{back_normalization}
\usage{
back_normalization(data)
}
\description{
A function for back_normalization.
}
\examples{
neuralnet::compute(model1,example1)
back_normalization(data=example1$net.result)
}
\keyword{deep}
\keyword{goat}
\keyword{learning}
\keyword{methane}
