% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_frame.R
\name{base_frame}
\alias{base_frame}
\title{base_frame}
\usage{
base_frame(model = 1)
}
\description{
This function make base frame for input data.
}
\examples{
base_frame(model=1)
base_frame(model=2)
}
\keyword{deep}
\keyword{goat}
\keyword{learning}
\keyword{methane}
