% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denormalization.R
\name{denormalization}
\alias{denormalization}
\title{denormalization}
\usage{
denormalization(data)
}
\description{
A function for back_normalization.
}
\examples{
neuralnet::compute(model1,example1)
denormalization(data=example1$net.result)
}
\keyword{deep}
\keyword{goat}
\keyword{learning}
\keyword{methane}
