% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{normalization}
\usage{
normalization(data, model)
}
\description{
A function normalization.
}
\examples{
normalization(data=example1, model1)
}
\keyword{deep}
\keyword{goat}
\keyword{learning}
\keyword{methane}
